Tabla de estadisticas y arreglos

+++ src/main/java/ar/edu/unlam/tallerweb1/servicios/ServicioExpulsion.java
@@ -5,7 +5,7 @@ import ar.edu.unlam.tallerweb1.modelo.Jugador;
 
 public interface ServicioExpulsion {
 	
-	public Double cantidadExpulsiones(Long id);
+	public Integer cantidadExpulsiones(Long id);
 	public Double promedioExpulsiones(Jugador jugador);
 	public Double valoracionPorExpulsiones(Jugador jugador);
 	public void guardarExpulsion(Expulsion expulsion);
